+++ packages/ts-newskit/src/assets/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { customToNewsKitIcon, styled } from 'newskit';
+import { East } from '@emotion-icons/material';
 import SearchIcon from './SearchIcon';
 import BurgerIcon from './BurgerIcon';
 import RoundedPlusIcon from './RoundedPlusIcon';
@@ -9,6 +10,7 @@ import TimesMasthead from './TimesMasthead';
 import SundayTimesMasthead from './SundayTimesMasthead';
 import ChevronRigthIcon from './ChevronRightIcon';
 import ArticlePlaceholder from './ArticlePlaceholder';
+import WaveBg from './WaveBg';
 
 const RoundedCloseIcon = styled(RoundedPlusIcon)`
   transform: rotate(45deg);
@@ -57,3 +59,10 @@ export const NewsKitArticlePlaceholder = customToNewsKitIcon(
   'NewsKitArticlePlaceholder',
   props => <ArticlePlaceholder {...props} />
 );
+export const NewsKitWaveBg = customToNewsKitIcon('NewskitWaveBg', props => (
+  <WaveBg {...props} />
+));
+
+export const NewskitIconEast = customToNewsKitIcon('NewskitIconEast', props => (
+  <East {...props} />
+));