+++ packages/ts-newskit/src/components/puzzles/date-card/types.ts
@@ -1,0 +1,6 @@
+export type DateCardItems = {
+  publishedAt: string;
+  status?: 'COMPLETE' | 'IN PROGRESS' | string;
+  gameLevel?: string;
+  url: string;
+};