+++ packages/ts-newskit/src/components/puzzles/flag/__tests__/index.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import '@testing-library/jest-dom';
 import { render } from '../../../utils/test-utils';
 import { PuzzlesFlag } from '../index';
+import '@testing-library/jest-dom';
 
 describe('Flag test', () => {
   it('should render a snapshot', () => {
@@ -13,10 +13,4 @@ describe('Flag test', () => {
     const { asFragment } = render(<PuzzlesFlag status="IN PROGRESS" />);
     expect(asFragment()).toMatchSnapshot();
   });
-
-  /* it('should render the component', () => {
-    const { getByText } = render(<Breadcrumb data={breadcrumbItems} />);
-    const getBreadcrumb = getByText('Tennis');
-    expect(getBreadcrumb).toBeInTheDocument();
-  }); */
 });