+++ packages/ts-newskit/src/components/slices/lead-story/index.tsx
@@ -5,18 +5,20 @@ import {
   Visible,
   Block,
   Headline,
-  LinkInline
+  LinkInline,
+  useTheme
 } from 'newskit';
 import React from 'react';
-import { StyledTextBlock, StyledTextStack, StyledInterviewee } from './styles';
+import { ColouredText } from '../shared-styles';
+import { StyledTextBlock, StyledTextStack } from './styles';
 
-interface LeadStoryPropsDataProp {
+export interface LeadStoryProps {
   headline: string;
-  color: string;
+  color?: string;
   readingTime: string;
   summary: string;
   bylines: string;
-  interviewee: string;
+  subHeadline?: string;
   caption: string;
   image: string;
   url: string;
@@ -28,21 +30,16 @@ export const LeadStory = ({
   readingTime,
   summary,
   bylines,
-  interviewee,
+  subHeadline,
   caption,
   image,
   url
-}: LeadStoryPropsDataProp) => {
+}: LeadStoryProps) => {
+  const theme = useTheme();
   return (
     <Stack flow={{ sm: 'vertical-center', md: 'horizontal-stretch' }}>
-      <StyledTextStack
-        marginInlineEnd={{ md: 'space040' }}
-        marginInline={{
-          xs: 'space045',
-          md: 'space000'
-        }}
-      >
-        <StyledInterviewee
+      <StyledTextStack marginInlineEnd={{ md: 'space040' }}>
+        <ColouredText
           $color={color}
           paddingBlockStart={{
             xs: 'space040',
@@ -52,8 +49,8 @@ export const LeadStory = ({
           paddingBlockEnd="space020"
           typographyPreset="newPreset090"
         >
-          {interviewee}
-        </StyledInterviewee>
+          {subHeadline}
+        </ColouredText>
         <Headline
           headingAs="h1"
           overrides={{
@@ -99,6 +96,10 @@ export const LeadStory = ({
       <Block
         paddingBlockStart={{
           xs: 'space040',
+          md: 'space000'
+        }}
+        marginInline={{
+          xs: `-${theme.spacePresets.space045}`,
           md: 'space000'
         }}
       >