+++ packages/ts-newskit/src/theme/times-web-light/style-presets/index.ts
@@ -7,9 +7,11 @@ import { menuStylePresets } from './menu';
 import { navStylePresets } from './navigation';
 import { pageStylePresets } from './page';
 import { secondaryNavigationStylePresets } from './secondary-nav';
+import { sliceStylePresets } from './slices';
 import { textStylePresets } from './text';
 import { flagStylePresets } from './flag';
 import { storyCardStylePresets } from './storyCard';
+import { accordionStylePresets } from './accordion';
 
 export const stylePresets = {
   ...articleListStylePresets,
@@ -22,6 +24,8 @@ export const stylePresets = {
   ...navStylePresets,
   ...secondaryNavigationStylePresets,
   ...leadStoryStylePresets,
+  ...sliceStylePresets,
   ...flagStylePresets,
+  ...accordionStylePresets,
   ...storyCardStylePresets
 };