+++ packages/ts-newskit/src/types/externs.d.ts
@@ -1,0 +1,3 @@
+declare module '@times-components/ts-components' {
+  export const sanitiseCopy: (copy: string = '', allowedTags: {}) => string;
+}