+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.40.8](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.40.7...@times-components/article-comments@0.40.8) (2023-04-24)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
+## [0.40.7](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.40.6...@times-components/article-comments@0.40.7) (2023-04-19)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
+## [0.40.6](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.40.5...@times-components/article-comments@0.40.6) (2023-04-18)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
+## [0.40.5](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.40.4...@times-components/article-comments@0.40.5) (2023-04-06)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
+## [0.40.4](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.40.3...@times-components/article-comments@0.40.4) (2023-04-04)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 ## [0.40.3](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.40.2...@times-components/article-comments@0.40.3) (2023-04-03)
 
 **Note:** Version bump only for package @times-components/article-comments